requirement.txt
asgiref==3.8.1
Django==5.0.4
django-rest-framework==0.1.0
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
